changes in auth_menu

+++ main.py
@@ -14,12 +14,17 @@ def auth_menu():
             if token:
                 return token
         elif choice == '2':
-            auth.signup()
+            token= auth.signup()
+            if token:
+                return token
         elif choice == '3':
             print("Exiting...")
             return None
         else:
             print("Invalid choice. Please try again.")
+            
+def get_valid_payload(token):
+    return auth.decode_jwt_token(token)
 
 
 def main():
@@ -27,13 +32,23 @@ def main():
     token = auth_menu()
     if not token:
         return
-                            
+
+                          
     while True:
+        payload = get_valid_payload(token)
+        if payload is None:
+            print("Session expired or invalid token. Please log in again.")
+            token = auth_menu()
+            if not token:
+                return
+            continue
+        is_admin = payload.get("role") == "admin"
+        
         print("\nCustomer Contact Book")
         print("\n1. Add Customer")
         print("2. View Customers")
         print("3. Search Customer")
-        print("4. Delete Customer")
+        print("4. Delete Customer" +(""if is_admin else "(admin only)"))
         print("5. Update Customer")
         print("6. Exit")
         choice = input("\nEnter your choice: ").strip()
@@ -44,7 +59,17 @@ def main():
         elif choice == '3':
             customers.search_customer()
         elif choice == '4':
-            customers.delete_customer()
+            payload = get_valid_payload(token)
+            if payload is None:
+                print("Session expired or invalid token. Please log in again.")
+                token = auth_menu()
+                if not token:
+                    return
+                continue
+            if payload.get("role") != "admin":
+                print("You do not have permission to delete customers. Admin access required.")
+            else:
+                customers.delete_customer()
         elif choice == '5':
             customers.update_customer()  
         elif choice == '6':
